refactor: relocate tanstack helpers out of lib into their layers
queryKeySerialize builds a stable toast dedupe key from a query key, so it
moves to common/infrastructure/tanstack/query-key. filterColumnByDate is the
predicate half of the date-filter table feature, so it moves beside that
compound's UI and state as filter-column-by-date. Rewrite consumers onto the
new paths. While relocating the filter, scope its bare eslint-disable to
no-explicit-any with a reason.
