+++ src/common/presentation/ui/table/compounds/date-filter/filter-column-by-date.ts
@@ -2,7 +2,9 @@ import type { Row } from '@tanstack/react-table';
 
 type DateFilter = { from?: string; to?: string };
 
-// eslint-disable-next-line
+// The TanStack Row generic needs an index signature to read row.original[columnId];
+// unknown would force a cast on new Date(...).
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function filterColumnByDate<T extends Record<string, any>>(
   row: Row<T>,
   columnId: string,
