+++ src/module/car/ownership/presentation/ui/sections/ownerships/use-ownerships.ts
@@ -5,8 +5,8 @@ import {
   getOwnerProfilesQueryOptions,
   getOwnershipsByCarIdQueryOptions,
 } from '@/car/ownership/infrastructure/tanstack/query/options';
+import { queryKeySerialize } from '@/common/infrastructure/tanstack/query-key';
 import { useToasts } from '@/common/presentation/hook/use-toasts';
-import { queryKeySerialize } from '@/lib/tanstack/utils';
 import { useSessionUser } from '@/user/presentation/hooks/use-session-user';
 
 interface UseOwnershipsSectionParams {
